requirements.txt
Flask
Flask-SQLAlchemy
qrcode
Pillow
pytz